% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BlusterParam-class}
\alias{BlusterParam-class}
\title{The BlusterParam class}
\description{
The BlusterParam class is a virtual base class controlling S4 dispatch in \code{\link{clusterRows}} and friends.
Concrete subclasses specify the choice of clustering algorithm,
while the slots of an instance of such a subclass represent the parameters for that algorithm.
}
\seealso{
\linkS4class{HclustParam}, \linkS4class{KmeansParam} and \linkS4class{SnnGraphParam} 
for some examples of concrete subclasses.
}
\author{
Aaron Lun
}
