% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowSOMParam.R
\docType{class}
\name{FlowSOMParam-class}
\alias{FlowSOMParam-class}
\alias{show,FlowSOMParam-method}
\alias{FlowSOMParam}
\alias{clusterRows,ANY,FlowSOMParam-method}
\title{FlowSOM-based clustering}
\usage{
FlowSOMParam(
  centers,
  dim.ratio = 1,
  rlen = 10,
  mst = 1,
  alpha = c(0.05, 0.01),
  radius = NULL,
  init = FALSE,
  initf = NULL,
  distf = 2
)

\S4method{clusterRows}{ANY,FlowSOMParam}(x, BLUSPARAM, full = FALSE)
}
\arguments{
\item{centers}{An integer scalar specifying the number of centers.
Alternatively, a function that takes the number of observations and returns the number of centers.}

\item{dim.ratio}{A positive numeric scalar in specifying how \code{centers} should be distributed between the \code{x} and \code{y} dimensions.
Defaults to equal distribution, i.e., both dimensions will be of length equal to the square root of \code{centers}.
Values above 1 will distribute more nodes to \code{x} while values below 1 will distribute mode nodes to \code{y}.}

\item{rlen, mst, alpha, radius, init, initf, distf}{Further arguments to pass to the \code{\link[FlowSOM]{SOM}} function in the \pkg{FlowSOM} package.}

\item{x}{A numeric matrix-like object where rows represent observations and columns represent variables.}

\item{BLUSPARAM}{A \linkS4class{FlowSOMParam} object.}

\item{full}{Logical scalar indicating whether the full mini-batch k-means statistics should be returned.}
}
\value{
The \code{FlowSOMParam} constructor will return a \linkS4class{FlowSOMParam} object with the specified parameters.

The \code{clusterRows} method will return a factor of length equal to \code{nrow(x)} containing the cluster assignments.
If \code{full=TRUE}, a list is returned with \code{clusters} (the factor, as above) and \code{objects};
the latter will contain the direct output of \code{\link[FlowSOM]{SOM}}.
}
\description{
Use the self-organizing map implementation in the \pkg{FlowSOM} package to cluster observations into the specified number of nodes.
Note that this requires the installation of the \pkg{FlowSOM} package.
}
\details{
This class usually requires the user to specify the number of clusters beforehand.
However, we can also allow the number of clusters to vary as a function of the number of observations.
The latter is occasionally useful, e.g., to allow the clustering to automatically become more granular for large datasets.

Note that the final number of clusters may not be exactly equal to \code{centers}, depending on how \code{dim.ratio} is specified.
For example, if \code{centers} is a perfect square and \code{dim.ratio=1}, we will get exactly the requested number of points.

To modify an existing FlowSOMParam object \code{x},
users can simply call \code{x[[i]]} or \code{x[[i]] <- value} where \code{i} is any argument used in the constructor.

For \code{radius}, a value of \code{NULL} means that the default argument in the \code{\link[FlowSOM]{SOM}} function signature is used.
This is are data-dependent and so cannot be specified during construction of the FlowSOMParam object.
\code{initf} has the same behavior.
}
\examples{
clusterRows(iris[,1:4], FlowSOMParam(centers=16))
clusterRows(iris[,1:4], FlowSOMParam(centers=12, dim.ratio=3/4))

}
\seealso{
\code{\link[FlowSOM]{SOM}} from the \pkg{FlowSOM} package, which does all of the heavy lifting.

\linkS4class{FixedNumberParam}, the parent of the FlowSOMParam class.
}
\author{
Aaron Lun
}
