% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluster-package.R
\docType{package}
\name{bluster-package}
\alias{bluster}
\alias{bluster-package}
\title{bluster: Clustering Algorithms for Bioconductor}
\description{
provides common clustering algorithms in an S4 framework that is amenable to extension.
    Implementations and wrappers are provided for hierarchical, k-means and graph-based clustering.
    Several utilities are also provided to compare and evaluate clustering results.
}
\author{
\strong{Maintainer}: Aaron Lun \email{infinite.monkeys.with.keyboards@gmail.com}

}
\keyword{internal}
