% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluster-package.R
\docType{package}
\name{bluster-package}
\alias{bluster}
\alias{bluster-package}
\title{bluster: Clustering Algorithms for Bioconductor}
\description{
Wraps common clustering algorithms in an easily extended S4 framework. Backends are implemented for hierarchical, k-means and graph-based clustering. Several utilities are also provided to compare and evaluate clustering results.
}
\author{
\strong{Maintainer}: Aaron Lun \email{infinite.monkeys.with.keyboards@gmail.com}

Other contributors:
\itemize{
  \item Stephanie Hicks [contributor]
  \item Basil Courbayre [contributor]
  \item Tuomas Borman [contributor]
  \item Leo Lahti [contributor]
}

}
\keyword{internal}
