% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMMParam.R
\docType{class}
\name{DMMParam-class}
\alias{DMMParam-class}
\alias{.defaultScalarArguments,DMMParam-method}
\alias{show,DMMParam-method}
\alias{DMMParam}
\alias{clusterRows,ANY,DMMParam-method}
\alias{bestDMMFit}
\title{DMM clustering}
\usage{
DMMParam(k = NULL, type = NULL, transposed = FALSE, seed = NULL)

\S4method{clusterRows}{ANY,DMMParam}(x, BLUSPARAM, full = FALSE)

bestDMMFit(dmm, type)
}
\arguments{
\item{k}{An integer indicating the number of clusters, or a list of integers defining the possible number of clusters to choose from.}

\item{type}{A boolean specifying}

\item{seed}{An integer specifying the seed to use when doing the DMM algorithm. Will be random if not set.}

\item{x}{A numeric matrix-like object where rows represent observations and columns represent variables.}

\item{BLUSPARAM}{A \linkS4class{KmeansParam} object.}

\item{full}{Logical scalar indicating whether the clustering statistics from both steps should be returned.}
}
\value{
The \code{DMMParam} constructor will return a \linkS4class{DMMParam} object with the specified parameters.

The \code{clusterRows} method will return a factor of length equal to \code{nrow(x)} containing the cluster assignments.
If \code{full=TRUE}, a list is returned with \code{clusters} (the factor, as above) and \code{objects} 
(a list containing \code{dmm}, the dmm object; \code{k}, the number of clusters; 
\code{prob}, the array of probabilities; and \code{seed} the seed for the dmm).
}
\description{
Run the \code{DMM} function on a distance matrix within \code{\link{clusterRows}}.
}
\details{
The DMM algorithm (see reference below) is commonly used in microbial ecology along with metagenomic & 16S rRNA count data. 
Holmes I, Harris K, Quince C. Dirichlet multinomial mixtures: generative models for microbial metagenomics. PLoS One. 2012;7(2):e30126. doi: 10.1371/journal.pone.0030126. Epub 2012 Feb 3. PMID: 22319561; PMCID: PMC3272020.

Because of this specificity, the best data we can use for this algorithm is microbiota read counts (for example the Twins data set in the DMM package).
To modify an existing DMMParam object \code{x},
users can simply call \code{x[[i]]} or \code{x[[i]] <- value} where \code{i} is any argument used in the constructor.

If \code{k=NULL}, the cluster search range will be 1:3.
The default for the fit function is \code{laplace}.
}
\examples{
clusterRows(iris[,1:4], DMMParam())
clusterRows(iris[,1:4], DMMParam(k=2))
clusterRows(iris[,1:4], DMMParam(k=1:3, type="laplace"))
\dontrun{
example(fl <- system.file(package="DirichletMultinomial", "extdata", "Twins.csv")
        count <- t(as.matrix(read.csv(fl, row.names=1)))
        clusterRows(counts, DMMParam(k=1:3, type="laplace")))
}

}
\author{
Basil Courbayre
}
