% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KmeansParam.R
\docType{class}
\name{KmeansParam-class}
\alias{KmeansParam-class}
\alias{KmeansParam}
\alias{clusterRows,ANY,KmeansParam-method}
\title{K-means clustering}
\usage{
KmeansParam(centers, ...)

\S4method{clusterRows}{ANY,KmeansParam}(x, BLUSPARAM)
}
\arguments{
\item{centers}{Integer scalar specifying the number of centers.}

\item{...}{Further arguments to pass to \code{\link{kmeans}}.}

\item{x}{A numeric matrix-like object where rows represent observations and columns represent variables.}

\item{BLUSPARAM}{A \linkS4class{KmeansParam} object.}
}
\value{
The \code{KmeansParam} constructor will return a \linkS4class{KmeansParam} object with the specified parameters.

The \code{clusterRows} method will return a factor of length equal to \code{nrow(x)} containing the cluster assignments.
}
\description{
Run the base \code{\link{kmeans}} function with the specified number of centers.
}
\author{
Aaron Lun
}
